requirements.txt
pyserial==3.5
PyYAML==6.0.1
pymodbus
six==1.15.0
requests==2.31.0
